tests/wpt/web-platform-tests/navigation-api/navigate-event/navigate-history-back-after-pushState.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  window.onload = () => t.step_timeout(() => {
    let start_length = history.length;
    let target_key = navigation.currentEntry.key;
    let target_id = navigation.currentEntry.id;
    history.pushState(1, "", "pushState.html");
    assert_equals(history.length, start_length + 1);

    navigation.onnavigate = t.step_func_done(e => {
      assert_equals(e.navigationType, "traverse");
      assert_false(e.cancelable);
      assert_true(e.canTransition);
      assert_false(e.userInitiated);
      assert_false(e.hashChange);
      assert_equals(e.downloadRequest, null);
      assert_equals(new URL(e.destination.url).hash, "");
      assert_true(e.destination.sameDocument);
      assert_equals(e.destination.key, target_key);
      assert_equals(e.destination.id, target_id);
      assert_equals(e.destination.index, 0);
      assert_equals(e.formData, null);
    });

    history.back();
  }, 0);
}, "history.back() fires the navigate event when reversing a pushState()");
</script>
